Resolve 8 conflicts across 5 files:
- engine.py: adopt main's jax.jit decorator, open_url for HLO dumps
- attention.py: keep both HEAD's ragged paged attention config + main's gated field
- tpus.py: adopt main's improved capacity_type normalization
- jax_utils.py: combine imports from both sides (process_allgather + AxisType)
- test_engine.py: keep HEAD's dataclasses import

+++ lib/haliax/src/haliax/nn/normalization.py
@@ -1,7 +1,6 @@
-# Copyright 2025 The Levanter Authors
+# Copyright The Levanter Authors
 #
 # SPDX-License-Identifier: Apache-2.0
-
 
 import dataclasses
 from abc import abstractmethod
@@ -114,9 +113,9 @@ class LayerNorm(LayerNormBase):
         out = out.astype(dtype)
 
         if self.weight is not None:
-            out = self.weight * out
+            out = self.weight.astype(out.dtype) * out
         if self.bias is not None:
-            out = out + self.bias
+            out = out + self.bias.astype(out.dtype)
         return out
 
 
@@ -134,9 +133,9 @@ class RmsNorm(LayerNormBase):
         out = out.astype(in_dtype)
 
         if self.weight is not None:
-            out = self.weight * out
+            out = self.weight.astype(out.dtype) * out
         if self.bias is not None:
-            out = out + self.bias
+            out = out + self.bias.astype(out.dtype)
         return out
